refactor: separate Jazz client and server features
