+++ crates/jazz/src/tools/public_schema_convert.rs
