+++ crates/jazz/src/tools/server/runtime_catalogue.rs
@@ -8,7 +8,8 @@ use crate::protocol::{LensOp as CoreLensOp, MigrationLens, TableLens};
 use crate::tools::public_api::types::{Schema, SchemaHash, TableName, Value};
 use crate::tools::schema_lens::{Lens, LensOp};
 
-use super::{ServerState, core_server_shell::ServerShellHandle, public_schema_convert};
+use super::{ServerState, core_server_shell::ServerShellHandle};
+use crate::tools::public_schema_convert;
 
 /// Publish newly admitted catalogue entries into the active runtime shell.
 ///
